requirements.txt
pandas==2.0.3
PyYAML==6.0.1
streamlit==1.27.2
streamlit-authenticator==0.2.3
plotly==5.9.0